app/requirements.txt
alembic==1.11.1
aioconsole==0.6.2
anthropic==0.8.1
beautifulsoup4==4.12.2
chromadb==0.4.13
edge-tts==6.1.9
emoji==2.9.0
fastapi[all]==0.103.2
faster_whisper==0.9.0
firebase_admin==6.2.0
google-cloud-speech==2.23.0
httpx==0.25.0
langchain==0.0.308
llama_index==0.8.39.post2
numpy==1.26.0
openai==0.27.2
pgvector==0.2.3
psycopg2-binary==2.9.9
pyaudio==0.2.13
pydantic==2.4.2
pydub==0.25.1
pypdf==3.16.2
pytest==7.4.2
python-dotenv==1.0.0
readerwriterlock==1.0.9
rebyte-langchain>=0.0.5
Requests==2.31.0
simpleaudio==1.0.4
SpeechRecognition==3.10.0
SQLAlchemy==2.0.21
starlette==0.27
twilio==8.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
transformers